Re-add auction results
